Add popover examples and fix bugs
Add a couple initial popover examples in the docs so people can play with it.
Add a fix for the scoped styles, remove the scoped from the style block inside the component
so people can style the component at will. Add a class at the top level of the component,
this class should be the required name passed in.

Update the tests to include a test for the top level class name. Change the face
to an ahref by default.